+++ resources/views/app.blade.php
@@ -6,6 +6,17 @@
 
         <title inertia>{{ config('app.name', 'Laravel') }}</title>
 
+        <!-- Favicon -->
+        @php
+            $systemSettings = \App\Models\Setting::values();
+            $favicon = $systemSettings['app_favicon'] ?? null;
+        @endphp
+        @if($favicon)
+            <link rel="icon" href="{{ $favicon }}">
+        @else
+            <link rel="icon" type="image/x-icon" href="{{ asset('favicon.ico') }}">
+        @endif
+
         <!-- Fonts -->
         <link rel="preconnect" href="https://fonts.googleapis.com">
         <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>